# src/main/resources/swagger.yaml
openapi: 3.0.3
info:
  title: Capgemini Bank
  description: A bank to manage accounts, including opening an account with an option to add initial credit, retrieving account balance, retrieving person information, and making transfers between accounts.
  version: 1.0.0
  contact:
    email: monali.aich@capgemini.com
    name: Monali Aich
servers:
  - url: http://api/capbank/v1
    description: Production server
  - url: http://localhost:{ports}/api/staging/capbank/v1
    description: Staging server
    variables:
      ports:
        enum:
          - '9000'
          - '7000'
        default: !!str 8080
  - url: http://localhost:8080/api/capbank/v1
    description: Local server     
tags:
- name: Customer
  description: Overview of the customer details.
- name: Account 
  description: Overview of the account details.
paths:   
  /account/saveAccntWithCrdtOrNoCrdt:
    post:
      tags:
        - Account
      summary: Add an account
      description: API to open an account with the option to add initial credit.
      operationId: addCustomer
      requestBody:
        $ref: '#/components/requestBodies/AccountRequests'
      responses:
        201:
          $ref: '#/components/responses/AccountCreationResponses'
        400:
          description: Incorrect Input
        409:
          description: Conflict
        default:
          description: Success.
  /account/getAccountInfoAndBalance/{accountId}:
    get:
      tags:
        - Account
      summary: Retrieve Account Details
      description: API to retrieve account information including account balance.
      parameters:
        - name: accountId
          required: true
          in: path
          schema:
            type: integer
      operationId: getAccountDtlsById
      responses:
        200:
          $ref: '#/components/responses/AccountRetrievalResponses'
        400:
          description: Incorrect Input
        409:
          description: Conflict
        default:
          description: Success.
  /customer/getCustomer/{customerId}:
    get:
      tags:
        - Customer
      summary: Retrieve Customer Details
      description: API to retrieve person information.
      parameters:
        - $ref: '#/components/parameters/customerId'
      operationId: getCustomerDtlsById
      responses:
        200:
          $ref: '#/components/responses/CustomerRetrievalResponses'
        400:
          description: Incorrect Input
        409:
          description: Conflict
        default:
          description: Success.
components:
  parameters:
    customerId:
      name: customerId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/CustomerId'
    accountId:
      name: accountId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/AccountId'
  requestBodies:
    AccountRequests:
      description: Account Details
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Account'
    CustomerRequest:
      description: Customer Details
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Customer'
  responses:
    AccountRetrievalResponses:
      description: Successfully Account details Fetched.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Account'
    CustomerRetrievalResponses:
      description: Successfully Person details Fetched
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Customer'
    AccountCreationResponses:
      description: Successfully Account Created
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AccountCreated'
  
  schemas:
    CustomerId:
      type: string
      description: Id of a Customer
    AccountId:
      type: string
      description: Id of an Account
    Customer:
      type: object
      description: Details of account holders
      properties:
        customerId:
          $ref: '#/components/schemas/CustomerId'
        firstName:
          $ref: '#/components/schemas/FirstName'
        lastName:
          $ref: '#/components/schemas/LastName'
    Account:
      type: object
      description: Details of Bank Account
      properties:
        accountId:
          $ref: '#/components/schemas/AccountId'
        accountName:
          $ref: '#/components/schemas/AccountName'
        typeOfAccount:
          $ref: '#/components/schemas/TypeOfAccount'
    FirstName:
      type: string
      description: First name of the account holders.
    LastName:
      type: string
      description: Last name of the account holders.
    AccountName:
      type: string
      description: Account Holder name.
    TypeOfAccount:
      type: string
      description: Saving or Current type of account.
    AccountCreated:
      type: object
      properties:
        accountId:
          $ref: '#/components/schemas/AccountId'